moved the gear icon bottom of the layer to the toolbar
