--- src/components/common/icons/ImportFile.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { SvgIcon } from '@material-ui/core';
-
-export function ImportFileIcon(props: any) {
-  return (
-    <SvgIcon {...props}>
-      <path d="M19.41,7.41l-4.83-4.83C14.21,2.21,13.7,2,13.17,2H6C4.9,2,4.01,2.9,4.01,4L4,20c0,1.1,0.89,2,1.99,2H18c1.1,0,2-0.9,2-2 V8.83C20,8.3,19.79,7.79,19.41,7.41z M14.8,15H13v3c0,0.55-0.45,1-1,1s-1-0.45-1-1v-3H9.21c-0.45,0-0.67-0.54-0.35-0.85l2.8-2.79 c0.2-0.19,0.51-0.19,0.71,0l2.79,2.79C15.46,14.46,15.24,15,14.8,15z M14,9c-0.55,0-1-0.45-1-1V3.5L18.5,9H14z" />
-    </SvgIcon>
-  );
-}
