+++ src/components/configure/keymapToolbar/KeymapToolbar.tsx
@@ -17,6 +17,8 @@ import { KeymapPdfGenerator } from '../../../services/pdf/KeymapPdfGenerator';
 import Keymap from '../keymap/Keymap';
 import LightingDialog from '../lighting/LightingDialog';
 import LayoutOptionPopover from '../layoutoption/LayoutOptionPopover.container';
+import { ImportFileIcon } from '../../common/icons/ImportFileIcon';
+import ImportDefDialog from '../importDef/ImportDefDialog';
 
 type OwnProp = {};
 
@@ -27,6 +29,7 @@ type KeymapMenuPropsType = OwnProp &
 type OwnKeymapMenuStateType = {
   openLightingDialog: boolean;
   layoutOptionPopoverPosition: { left: number; top: number } | null;
+  openImportDefDialog: boolean;
 };
 
 export default class KeymapMenu extends React.Component<
@@ -38,6 +41,7 @@ export default class KeymapMenu extends React.Component<
     this.state = {
       openLightingDialog: false,
       layoutOptionPopoverPosition: null,
+      openImportDefDialog: false,
     };
   }
 
@@ -73,6 +77,14 @@ export default class KeymapMenu extends React.Component<
     this.setState({ layoutOptionPopoverPosition: null });
   }
 
+  private onClickOpenImportDefFileDialog() {
+    this.setState({ openImportDefDialog: true });
+  }
+
+  private onCloseImportDefFileDialog() {
+    this.setState({ openImportDefDialog: false });
+  }
+
   private onClickGetCheatsheet() {
     const keymaps: { [pos: string]: IKeymap }[] = this.props.keymaps!;
     const keys: { [pos: string]: Key }[] = [];
@@ -105,6 +117,11 @@ export default class KeymapMenu extends React.Component<
       this.props.keyboardDefinition!.lighting
     );
     const hasLayoutOptions = 0 < this.props.selectedKeyboardOptions!.length;
+    const {
+      vendorId,
+      productId,
+      productName,
+    } = this.props.keyboard!.getInformation();
     return (
       <React.Fragment>
         <div className="keymap-menu">
@@ -160,6 +177,28 @@ export default class KeymapMenu extends React.Component<
               />
             </div>
           )}
+
+          <div className="keymap-menu-item">
+            <Tooltip
+              arrow={true}
+              placement="top"
+              title="Import local keyboard definition file(.json)"
+            >
+              <IconButton
+                size="small"
+                onClick={this.onClickOpenImportDefFileDialog.bind(this)}
+              >
+                <ImportFileIcon />
+              </IconButton>
+            </Tooltip>
+            <ImportDefDialog
+              open={this.state.openImportDefDialog}
+              onClose={this.onCloseImportDefFileDialog.bind(this)}
+              vendorId={vendorId}
+              productId={productId}
+              productName={productName}
+            />
+          </div>
 
           <div className="keymap-menu-item">
             <Tooltip
